+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # Changelog
 
+## [v2.0.1](https://github.com/containeroo/helminator/tree/v2.0.1) (2020-12-11)
+
+[All Commits](https://github.com/containeroo/helminator/compare/v1.7.1...v2.0.1)
+
+**New features:**
+
+- Slack message contains a link to the related merge request
+- new variables:
+  - `HELMINATOR_GITLAB_REMOVE_SOURCE_BRANCH` -> Delete source branch when merge request is accepted
+  - `HELMINATOR_GITLAB_SQUASH_COMMITS` -> Squash commits when merge request is accepted
+
+**Bug fixes:**
+
+- various improvements
+
+### Changes from v2.0.0
+
+**New features:**
+
+- Introducing pull requests:
+  - create a pull request when an update is found
+  - assign pull request to one or more users
+  - update existing pull requests when a new update is found
+  - skip updating by closing the pull request
+- Added the ability to disable SSL verification (e.g. when behind a corporate proxy)
+
+**Dependencies:**
+
+- Update requests to v2.25.0 (#42)
+- Update semver to v2.14.0 (#40)
+- Update slackclient to v2.9.3 (#41)
+
+**Bug fixes:**
+
+- minor bugfixes
+- typos
+
 ## [v1.7.1](https://github.com/containeroo/helminator/tree/v1.7.1) (2020-10-12)
 
 [All Commits](https://github.com/containeroo/helminator/compare/v1.7.0...v1.7.1)